🧠Refactor(userService.ts): 함수 성격에 따른 파일 분리 및 함수 재배치   1. **auth.service.ts** - 인증과 관련된 함수들     - signUp     - signIn     - resetPassword   2. **user.service.ts** - 사용자 정보 변경과 관련된 기본 함수들     - updateUserInfo     - deleteUser   3. **userContent.service.ts** - 사용자 콘텐츠 관련 함수들     - findUserInfoByUserId     - findUserFeedsByUserId     - findUserCommentsByUserId     - findUserFeedSymbolsByUserId   4. **validator.service.ts** - DB 입력시 유효성 검사와 관련된 함수들     - checkDuplicateNickname     - checkDuplicateEmail
