+++ src/controllers/users/userContent.controller.ts
@@ -1,0 +1,90 @@
+// 유저의 가입정보 가져오기
+import userContentService from '../../services/users/userContent.service';
+import { Request, Response } from 'express';
+import { Pagination } from '../../repositories/feed.repository';
+
+const getUserInfo = async (req: Request, res: Response): Promise<void> => {
+  let targetUserId = Number(req.params.id);
+  const loggedInUserId = req.userInfo.id;
+
+  if (!targetUserId) {
+    targetUserId = loggedInUserId;
+  }
+  const result = await userContentService.findUserInfoByUserId(targetUserId);
+
+  res.status(200).json(result);
+};
+
+// 유저의 모든 게시물 가져오기
+const getUserFeeds = async (req: Request, res: Response): Promise<void> => {
+  let targetUserId = Number(req.params.id);
+  const loggedInUserId = req.userInfo.id;
+
+  if (!targetUserId) {
+    targetUserId = loggedInUserId;
+  }
+
+  const startIndex = Number(req.query.page);
+  const limit = Number(req.query.limit);
+  const page: Pagination = { startIndex, limit };
+
+  const result = await userContentService.findUserFeedsByUserId(
+    targetUserId,
+    page
+  );
+
+  res.status(200).json(result);
+};
+
+// 유저의 모든 덧글 가져오기
+const getUserComments = async (req: Request, res: Response): Promise<void> => {
+  let targetUserId = Number(req.params.id);
+  const loggedInUserId = req.userInfo.id;
+
+  if (!targetUserId) {
+    targetUserId = loggedInUserId;
+  }
+
+  const startIndex = Number(req.query.index);
+  const limit = Number(req.query.limit);
+  const page: Pagination = { startIndex, limit };
+
+  const result = await userContentService.findUserCommentsByUserId(
+    targetUserId,
+    loggedInUserId,
+    page
+  );
+
+  res.status(200).json(result);
+};
+
+// 유저의 모든 좋아요 가져오기
+const getUserFeedSymbols = async (
+  req: Request,
+  res: Response
+): Promise<void> => {
+  let targetUserId = Number(req.params.id);
+  const loggedInUserId = req.userInfo.id;
+
+  if (!targetUserId) {
+    targetUserId = loggedInUserId;
+  }
+
+  const startIndex = Number(req.query.page);
+  const limit = Number(req.query.limit);
+  const page: Pagination = { startIndex, limit };
+
+  const result = await userContentService.findUserFeedSymbolsByUserId(
+    targetUserId,
+    page
+  );
+
+  res.status(200).json(result);
+};
+
+export default {
+  getUserInfo,
+  getUserFeeds,
+  getUserComments,
+  getUserFeedSymbols,
+};
